% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Example Multimodal Data}
\format{
A list with components:
\describe{
  \item{m1}{A matrix (300 x 1000) for modality 1}
  \item{m2}{A matrix (300 x 1000) for modality 2}
  \item{distmat}{A 1000 x 1000 distance matrix}
  \item{x}{A numeric vector of age}
  \item{z}{A factor or numeric vector of sex}
}
}
\usage{
data(Data)
}
\description{
A simulated or real dataset for demonstrating the use of `Ceidr`.
}
\keyword{datasets}
